finished metapage and started on btree logic

+++ Database/src/btree/node.cpp
@@ -1,10 +1,10 @@
 #include "../../include/btree/node.hpp"
 #include <cstring>
 
-constexpr size_t NODE_TYPE_OFFSET = 0;
-constexpr size_t NODE_TYPE_SIZE = 1;
-constexpr size_t NUM_KEYS_OFFSET = NODE_TYPE_OFFSET + NODE_TYPE_SIZE;
-constexpr size_t NUM_KEYS_SIZE = 4;
+constexpr uint64_t NODE_TYPE_OFFSET = 0;
+constexpr uint64_t NODE_TYPE_SIZE = 1;
+constexpr uint64_t NUM_KEYS_OFFSET = NODE_TYPE_OFFSET + NODE_TYPE_SIZE;
+constexpr uint64_t NUM_KEYS_SIZE = 4;
 
 NodeType getNodeType(Page* page){
     return static_cast<NodeType>(page->data[NODE_TYPE_OFFSET]);
